requirements.txt
streamlit
pandas
altair
pywebview
